Introduce new type "PortId" to prepare driver implementation.
Create a common base class for processes and ports.
In Erlang docs, "port" means both the external process and its identifier.
To distinguish both, the identifier is named "PortId" and the external
process is named "Port". The base class for processes and ports is therefore
named "ProcessOrPort" and the base class for the identifiers is named
"PidOrPortId". Maybe the class "PID" will be renamed to "Pid" later to
have a consistent naming.